LayoutTests/inspector/unit-tests/string-utilities-expected.txt


== Running test suite: StringUtilities
-- Running test case: String.format
PASS: float format specifier with no sub-specifier should show 6 decimal digits
PASS: float format specifier with precision 0 should show 0 decimal digits
PASS: float format specifier with precision 1 should show 1 decimal digit
PASS: float format specifier with precision 2 should show 2 decimal digits
PASS: float format specifier with precision 3 should show 3 decimal digits
PASS: float format specifier with precision 4 should show 4 decimal digits
PASS: float format specifier with precision 5 should show 5 decimal digits
PASS: float format specifier with precision 6 should show 6 decimal digits
PASS: float format specifier with precision 7 should show 7 decimal digits
PASS: float format specifier with precision 8 should show 8 decimal digits
PASS: float format specifier with precision 9 should show 9 decimal digits
PASS: float format specifier with string argument should attempt conversion to float
PASS: float format specifier with Infinity argument should show "∞"
PASS: float format specifier with NaN argument should show "NaN"
PASS: integer format specifier with float argument should convert to integer
PASS: integer format specifier with string argument should attempt conversion to integer
PASS: integer format specifier with Infinity argument should show "NaN"
PASS: integer format specifier with NaN argument should show "NaN"

-- Running test case: String.prototype.extendedLocaleCompare
PASS: "1" < "2"
PASS: "2" > "1"
PASS: "2" < "10"
PASS: "10" > "2"
PASS: "1" < "10"
PASS: "10" > "1"
PASS: "a1" < "a2"
PASS: "a2" > "a1"
PASS: "a2" < "a10"
PASS: "a10" > "a2"
PASS: "a1" < "a10"
PASS: "a10" > "a1"

-- Running test case: String.prototype.get lineCount
PASS: A string with two line breaks should have three lines.
PASS: A string with two consecutive line breaks should have three lines.
PASS: A string with a traling line breaks should have two lines.
PASS: An empty string is one line.

-- Running test case: String.prototype.get lastLine
PASS: Last line of one line string is the same string.
PASS: Last line of a three line string should be the third line.
PASS: Last line of a string with a traling line break should be empty.
PASS: Last line of an empty string is the same empty string.

-- Running test case: String.prototype.truncateStart
PASS: String stays the same.
PASS: Ellipsis is inserted before the third character.
PASS: Ellipsis is inserted before the fourth character.

-- Running test case: String.prototype.truncateMiddle
PASS: String stays the same.
PASS: Ellipsis is inserted in the middle.
PASS: Ellipsis is inserted after the second character.

-- Running test case: String.prototype.truncateEnd
PASS: String stays the same.
PASS: Ellipsis is inserted after the fourth character.
PASS: Ellipsis is inserted after the third character.

-- Running test case: String.prototype.escapeCharacters
PASS: String stays the same with no escape characters.
PASS: String stays the same with empty escape characters.
PASS: String stays the same with no matching escape characters.
PASS: The letter 'c' is escaped.
PASS: The letter 'c' and 'e' are escaped.
PASS: The letter 'c' and 'e' are escaped.
PASS: The letter 'c', 'd', and 'e' are escaped.
